Split json into parse and stringify
